Hygiene: input validation, enum-switch dispatch, stricter warnings, sanitizer preset
- demod: reject --slowdown <= 0 (the float->uint32 conversion of the
  resulting inf was undefined behaviour)
- sweeps: validate pll_ki >= 0 (a negative ki inverts the integral action
  into a confusing drift-to-the-rail failure)
- fir: notch_kernel enforces its documented odd-taps precondition instead
  of silently building a mis-centred filter
- chroma: validate the config before any of the four FIR kernels is built;
  dispatch CombMode by switch (no default), so a new mode fails to compile
  rather than silently decoding as PAL-S. Bit-identical output (verified
  against main across all three comb modes on corpus/wb3)
- one declaration per line in biquad.hpp, chroma_decoder.hpp,
  render_command.hpp, sync_command.cpp, MixerTest.cpp
- demod.cpp: derive kTwoPi from std::numbers::pi; IWYU <numbers>; drop
  unused <complex> from render_command.cpp / sync_command.cpp
- pedantic.cmake: add -Wdouble-promotion (mechanises the float/double
  rule), -Wnull-dereference, -Wnon-virtual-dtor, -Wsuggest-override, and
  the GCC-only -Wduplicated-cond/-branches and -Wlogical-op; fix the
  handful of implicit float->double promotions it found in tests and the
  kernel normalisation loops
- presets: -D_GLIBCXX_ASSERTIONS in debug; new "sanitize" preset
  (ASan+UBSan, serial PSTL backend so stdexec's legacy <execution>
  include doesn't drag in a TBB link dependency); full test suite and a
  colour corpus render are clean under it
- README: drop the colour-registration backlog item (fixed in #12) and
  the demod --no-sound-trap/--sound-q flags (removed with the sound trap)

+++ lib/test/DemodTest.cpp
@@ -81,7 +81,7 @@ TEST_CASE("Hilbert forms a one-sided analytic signal of the right sign") {
   double phase_step = 0.0;
   for (std::size_t k = lo; k < hi; ++k) {
     CHECK_THAT(std::abs(out[k]), WithinAbs(1.0, 0.03)); // one-sided => |.| ~ amplitude
-    phase_step += std::arg(out[k] * std::conj(out[k - 1]));
+    phase_step += static_cast<double>(std::arg(out[k] * std::conj(out[k - 1])));
   }
   const double mean_step = phase_step / static_cast<double>(hi - lo - 1);
   CHECK_THAT(mean_step, WithinAbs(two_pi * f / fs, 0.02)); // positive, ~ +2*pi*f/fs
